refact: Extract reusable TestLifecycle
